Working on comments on PR: Feature/smokescreen compartment 1 #6 regarding Oracle connetion Also removing duplicated lines from the active batch list
