+++ pages/batch_list_page.py
@@ -1,11 +1,11 @@
-from playwright.sync_api import Page
+from playwright.sync_api import Page, expect
 from utils.click_helper import click
 
 
-class ActiveBatchList:
+class BatchList:
     def __init__(self, page: Page):
         self.page = page
-        # Active Batch List - page filters
+        #Batch List - page filters
         self.id_filter = self.page.locator("#batchIdFilter")
         self.type_filter = self.page.locator("#batchTypeFilter")
         self.original_filter = self.page.locator("#originalBatchIdFilter")
@@ -14,44 +14,48 @@ class ActiveBatchList:
         self.batch_split_by_filter = self.page.locator("#splitByFilter")
         self.screening_centre_filter = self.page.locator("#screeningCentreFilter")
         self.count_filter = self.page.locator("#countFilter")
+        self.table_data = self.page.locator("td")
         self.batch_successfully_archived_msg = self.page.locator('text="Batch Successfully Archived and Printed"')
 
+    def verify_table_data(self, value):
+        expect(self.table_data.filter(has_text=value)).to_be_visible()
+
     def enter_id_filter(self, search_text: str):
-        click(self.page, self.id_filter)
         self.id_filter.fill(search_text)
         self.id_filter.press("Enter")
 
     def enter_type_filter(self, search_text: str):
-        click(self.page, self.type_filter)
         self.type_filter.fill(search_text)
         self.type_filter.press("Enter")
 
     def enter_original_filter(self, search_text: str):
-        click(self.page, self.original_filter)
         self.original_filter.fill(search_text)
         self.original_filter.press("Enter")
 
     def enter_event_code_filter(self, search_text: str):
-        click(self.page, self.event_code_filter)
         self.event_code_filter.fill(search_text)
         self.event_code_filter.press("Enter")
 
     def enter_description_filter(self, search_text: str):
-        click(self.page, self.description_filter)
         self.description_filter.fill(search_text)
         self.description_filter.press("Enter")
 
     def enter_batch_split_by_filter(self, search_text: str):
-        click(self.page, self.batch_split_by_filter)
         self.batch_split_by_filter.fill(search_text)
         self.batch_split_by_filter.press("Enter")
 
     def enter_screening_centre_filter(self, search_text: str):
-        click(self.page, self.screening_centre_filter)
         self.screening_centre_filter.fill(search_text)
         self.screening_centre_filter.press("Enter")
 
     def enter_count_filter(self, search_text: str):
-        click(self.page, self.count_filter)
         self.count_filter.fill(search_text)
         self.count_filter.press("Enter")
+
+class ActiveBatchList(BatchList):
+    def __init__(self, page):
+        super().__init__(page)
+
+class ArchivedBatchList(BatchList):
+    def __init__(self, page):
+        super().__init__(page)
