+++ pages/log_devices_page.py
@@ -12,12 +12,11 @@ class LogDevices:
         self.save_and_log_device_button = self.page.get_by_role("button", name="Save and Log Device")
         self.device_spoilt_button = self.page.get_by_role("button", name="Device Spoilt")
         self.sample_date_field = self.page.locator("#sampleDate")
-        self.successfully_logged_device_text = self.page.get_by_text("Successfully logged device")
+        self.successfully_logged_device_text = self.page.get_by_text("×Successfully logged device")
         self.spoilt_device_dropdown = self.page.get_by_label("Spoil reason drop down")
         self.log_as_spoilt_button = self.page.get_by_role("button", name="Log as Spoilt")
 
     def fill_fit_device_id_field(self, value):
-        click(self.page, self.fit_device_id_field)
         self.fit_device_id_field.fill(value)
         self.fit_device_id_field.press("Enter")
 
@@ -28,7 +27,6 @@ class LogDevices:
         click(self.page, self.device_spoilt_button)
 
     def fill_sample_date_field(self, value):
-        click(self.page, self.sample_date_field)
         self.sample_date_field.fill(value)
         self.sample_date_field.press("Enter")
