+++ tests/test_home_page_links.py
@@ -52,8 +52,7 @@ def test_home_page_links_navigation(page: Page) -> None:
     # Click the refresh alerts link
     homepage.click_refresh_alerts_link()
     # Verify that the 'last updated' timestamp matches the current date and time
-    (expect(page.locator("form[name=\"refreshCockpit\"]")).to_contain_text
-     ("Refresh alerts (last updated :" + DateTimeUtils.current_datetime()))
+    (expect(page.locator("form[name=\"refreshCockpit\"]")).to_contain_text("Refresh alerts (last updated :" + DateTimeUtils.current_datetime()))
 
     # Click the user guide link
     with page.expect_popup() as page1_info:
