--- utils/get_nhs_no_from_batch_id.py
@@ -1,16 +1,0 @@
-import pandas as pd
-from oracle import OracleDB
-
-
-def get_nhs_no_from_batch_id(batch_id):
-    nhs_number_df = OracleDB().execute_query(f"""
-    SELECT SUBJECT_NHS_NUMBER
-    FROM SCREENING_SUBJECT_T ss
-    INNER JOIN sd_contact_t c ON ss.subject_nhs_number = c.nhs_number
-    INNER JOIN LETT_BATCH_RECORDS lbr
-    ON ss.SCREENING_SUBJECT_ID = lbr.SCREENING_SUBJECT_ID
-    WHERE lbr.BATCH_ID IN {batch_id}
-    AND ss.screening_status_id != 4008
-    ORDER BY ss.subject_nhs_number
-    """)
-    return nhs_number_df
