--- utils/oracle.py
@@ -1,146 +1,0 @@
-import oracledb
-import os
-from dotenv import load_dotenv
-from sqlalchemy import create_engine
-import pandas as pd
-import logging
-
-
-class OracleDB:
-    def __init__(self):
-        load_dotenv()
-        self.user = os.getenv("un")
-        self.dns = os.getenv("cs")
-        self.password = os.getenv("pw")
-
-    def exec_bcss_timed_events(self, nhs_number_df):  # Executes bcss_timed_events when given NHS numbers
-        try:
-            logging.info("Attempting DB connection...")
-            conn = oracledb.connect(user=self.user, password=self.password, dsn=self.dns)
-            logging.info("DB connection successful!")
-            try:
-                for index, row in nhs_number_df.iterrows():
-                    logging.info(f"Attempting to get subject_id from nhs number: {row["subject_nhs_number"]}")
-                    cursor = conn.cursor()
-                    cursor.execute(
-                        f"SELECT SCREENING_SUBJECT_ID FROM SCREENING_SUBJECT_T WHERE SUBJECT_NHS_NUMBER = {int(row["subject_nhs_number"])}")
-                    result = cursor.fetchall()
-                    subject_id = result[0][0]
-                    logging.info(f"Able to extract subject ID: {subject_id}")
-                    try:
-                        logging.info(
-                            f"Attempting to execute stored procedure: {f"'bcss_timed_events', [{subject_id},'Y']"}")
-                        cursor = conn.cursor()
-                        cursor.callproc('bcss_timed_events', [subject_id, 'Y'])
-                        logging.info("Stored procedure execution successful!")
-                    except Exception as spExecutionError:
-                        logging.error(f"Failed to execute stored procedure with execution error: {spExecutionError}")
-            except Exception as queryExecutionError:
-                logging.error(f"Failed to to extract subject ID with error: {queryExecutionError}")
-            finally:
-                if conn is not None:
-                    conn.close()
-                    logging.info("Connection closed")
-        except Exception as connectionError:
-            logging.error(f"Failed to connect to the DB! with connection error: {connectionError}")
-
-    def populate_ui_approved_users_table(self, user: str):  # To add users to the UI_APPROVED_USERS table
-        try:
-            logging.info("Attempting DB connection...")
-            conn = oracledb.connect(user=self.user, password=self.password, dsn=self.dns)
-            logging.info("DB connection successful!")
-            try:
-                logging.info("Attempting to write to the db...")
-                cursor = conn.cursor()
-                cursor.execute(f"INSERT INTO UI_APPROVED_USERS (OE_USER_CODE) VALUES ('{user}')")
-                conn.commit()
-                logging.info("DB write successful!")
-            except Exception as dbWriteError:
-                logging.error(f"Failed to write to the DB! with write error {dbWriteError}")
-            finally:
-                if conn is not None:
-                    conn.close()
-                    logging.info("Connection closed")
-        except Exception as connectionError:
-            logging.error(f"Failed to connect to the DB! with connection error {connectionError}")
-
-    def delete_all_users_from_approved_users_table(self):  # To remove all users from the UI_APPROVED_USERS table
-        try:
-            logging.info("Attempting DB connection...")
-            conn = oracledb.connect(user=self.user, password=self.password, dsn=self.dns)
-            logging.info("DB connection successful!")
-            try:
-                logging.info("Attempting to delete users from DB table...")
-                cursor = conn.cursor()
-                cursor.execute(f"DELETE FROM UI_APPROVED_USERS WHERE OE_USER_CODE is not null")
-                conn.commit()
-                logging.info("DB table values successfully deleted!")
-            except Exception as dbValuesDeleteError:
-                logging.error(
-                    f"Failed to delete values from the DB table! with data deletion error {dbValuesDeleteError}")
-            finally:
-                if conn is not None:
-                    conn.close()
-                    logging.info("Connection Closed")
-        except Exception as connectionError:
-            logging.error(f"Failed to connect to the DB! with connection error {connectionError}")
-
-    def execute_query(self, query: str, ):  # To use when "select xxxx" (stored procedures)
-        try:
-            logging.info("Attempting DB connection...")
-            conn = oracledb.connect(user=self.user, password=self.password, dsn=self.dns)
-            engine = create_engine('oracle+oracledb://', creator=lambda: conn)
-            logging.info("DB connection successful!")
-            try:
-                logging.info(f"Attempting to execute query")
-                df = pd.read_sql(query, engine)
-                logging.info("query execution successful!")
-            except Exception as executionError:
-                logging.error(f"Failed to execute query with execution error {executionError}")
-            finally:
-                if conn is not None:
-                    conn.close()
-                    logging.info("Connection Closed - Returning results")
-                    return df
-        except Exception as connectionError:
-            logging.error(f"Failed to connect to the DB! with connection error {connectionError}")
-
-    def execute_stored_procedure(self, procedure: str):  # To use when "exec xxxx" (stored procedures)
-        try:
-            logging.info("Attempting DB connection...")
-            conn = oracledb.connect(user=self.user, password=self.password, dsn=self.dns)
-            logging.info("DB connection successful!")
-            try:
-                logging.info(f"Attempting to execute stored procedure: {procedure}")
-                cursor = conn.cursor()
-                cursor.callproc(procedure)
-                conn.commit()
-                logging.info("stored procedure execution successful!")
-            except Exception as executionError:
-                logging.error(f"Failed to execute stored procedure with execution error: {executionError}")
-            finally:
-                if conn is not None:
-                    conn.close()
-                    logging.info("Connection Closed")
-        except Exception as connectionError:
-            logging.error(f"Failed to connect to the DB! with connection error: {connectionError}")
-
-    def update_or_insert_data_to_table(self, statement, params):  # To update or insert data into a table
-        try:
-            logging.info("Attempting DB connection...")
-            conn = oracledb.connect(user=self.user, password=self.password, dsn=self.dns)
-            logging.info("DB connection successful!")
-            try:
-                logging.info("Attempting to insert/update table")
-                cursor = conn.cursor()
-                cursor.execute(statement, params)
-                conn.commit()
-                logging.info("DB table successfully updated!")
-            except Exception as dbUpdateInsertError:
-                logging.error(f"Failed to insert/update values from the DB table! with error {dbUpdateInsertError}")
-            finally:
-                if conn is not None:
-                    conn.close()
-                    logging.info("Connection Closed")
-        except Exception as connectionError:
-            logging.error(f"Failed to connect to the DB! with connection error {connectionError}")
